Factor platform specific memory pressure handling out of arc
